+++ src/DB/SQLiteFile.cpp
@@ -87,7 +87,7 @@ DB::DBConn::ConnType DB::SQLiteFile::GetConnType()
 	return CT_SQLITE;
 }
 
-Int32 DB::SQLiteFile::GetTzQhr()
+Int8 DB::SQLiteFile::GetTzQhr()
 {
 	return 0;
 }
